Added support for creating and receiving offers.
Added a framework for dealing with offers. A team has scaffolded ability
to determine offers it wants to send, and then sending those to players.
A player can receive offers, and has the ability scaffolded to evaluate
offers and accept the best offer if it deems it good enough. Specific
utility values are needed for this to be complete.