Implement Salary Cap for Teams.
Resolves #9. Adds a salary cap for each team, which is the true salary
cap for the NBA in the 2014-15 season. Teams will be aware of how much
money they have spent in a year on contracts, and will eventually not be
allowed to offer any contracts that would put them over the salary cap.
Additional work will also allow for a team to potentially exceed the
salary cap, at an additional penalty to the team.